Replace example files with testable examples
